requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==3.1.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
pycparser==2.19
python-decouple==3.1
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.3
